refactor(codec): extract peek emitters into CodecEmitterPeek.kt (step 5)
Move the peekFrameSize emit cluster out of CodecEmitter into a new
package-internal top-level file: buildPeekFrameFun and its specialized
walkers (appendPeekUseCodecScalar / appendPeekBoundingDynamicPrior /
appendPeekVariableLengthUseCodecScalar / appendPeekLengthPrefixedUseCodecList
/ appendSequentialPeek and its appendSequentialPeek* / appendPeek* helpers),
the peekBudgetFor heuristic, and the isPeekCollapsingConditional predicate.

Two stateless helpers shared with the decode visitor (decodeConditionExpr,
appendLengthFromIntMaxGuard) also move so the top-level peek functions can
call them; CodecEmitter's decode code calls them unqualified across the
package — the same cross-file pattern step 4 used for decodeAccessor /
appendPeekFixedScalar in CodecEmitterDispatch.kt.

All moved verbatim (no batchCounter/codeGenerator/logger state); CodecEmitter
5058 -> 3862 lines. Byte-identical codegen verified by CodecSnapshotTest;
processor + codec-test jvmTest suites green.
